+++ apps/api/src/original-torrent-session/index.ts
@@ -1,1 +1,5 @@
-export { OriginalTorrentSessionModule } from './session.module';
+export {
+  ORIGINAL_TORRENT_SESSION_CONFIG,
+  OriginalTorrentSessionModule,
+} from './session.module';
+export type { OriginalTorrentSessionConfig } from './session.config';